chore: move apps into old app structures
