--- packages/app-store/hubspot-revert/DESCRIPTION.md
@@ -1,11 +1,0 @@
----
-items:
-  - hubspot01.webp
----
-
-HubSpot is a cloud-based CRM designed to help align sales and marketing teams, foster sales enablement, boost ROI and optimize your inbound marketing strategy to generate more, qualified leads.
-
-Features:
-
-- Creates a meeting in HubSpot for one or many attendees
-- Attendees that don't exist in HubSpot, gets created first to associate them in the meeting
