--- packages/app-store/hubspot-revert/api/index.ts
@@ -1,2 +1,0 @@
-export { default as add } from "./add";
-export { default as callback } from "./callback";
