--- packages/app-store/hubspot-revert/lib/index.ts
@@ -1,1 +1,0 @@
-export { default as CalendarService } from "./CalendarService";
