--- packages/app-store/hubspot-revert/zod.ts
@@ -1,8 +1,0 @@
-import { z } from "zod";
-
-export const appKeysSchema = z.object({
-  client_id: z.string().min(1),
-  client_secret: z.string().min(1),
-});
-
-export const appDataSchema = z.object({});
