--- packages/app-store/hubspot/_metadata.ts
@@ -1,22 +1,0 @@
-import type { AppMeta } from "@calcom/types/App";
-
-import _package from "./package.json";
-
-export const metadata = {
-  name: "HubSpot CRM",
-  installed: !!process.env.HUBSPOT_CLIENT_ID,
-  description: _package.description,
-  type: "hubspot_crm",
-  variant: "crm",
-  logo: "icon.svg",
-  publisher: "Cal.com",
-  url: "https://hubspot.com/",
-  categories: ["crm"],
-  label: "HubSpot CRM",
-  slug: "hubspot",
-  title: "HubSpot CRM",
-  email: "help@cal.com",
-  dirName: "hubspot",
-} as AppMeta;
-
-export default metadata;
