+++ packages/app-store/hubspot/config.json
@@ -1,17 +1,17 @@
 {
   "/*": "Don't modify slug - If required, do it using cli edit command",
   "name": "Hubspot CRM",
-  "slug": "hubspot-revert",
-  "type": "hubspot-revert_other_calendar",
+  "slug": "hubspot",
+  "type": "hubspot_other_calendar",
   "logo": "icon.svg",
   "url": "https://revert.dev",
   "variant": "crm",
   "categories": ["crm"],
   "publisher": "Revert.dev ",
-  "email": "jatin@revert.dev",
+  "email": "support@revert.dev",
   "description": "HubSpot is a cloud-based CRM designed to help align sales and marketing teams, foster sales enablement, boost ROI and optimize your inbound marketing strategy to generate more, qualified leads.",
   "isTemplate": false,
   "__createdUsingCli": true,
   "__template": "basic",
-  "dirName": "hubspot-revert"
+  "dirName": "hubspot"
 }
