+++ packages/app-store/hubspot/index.ts
@@ -1,3 +1,2 @@
 export * as api from "./api";
 export * as lib from "./lib";
-export { metadata } from "./_metadata";
