+++ packages/app-store/index.ts
@@ -9,7 +9,6 @@ const appStore = {
   googlecalendar: () => import("./googlecalendar"),
   googlevideo: () => import("./googlevideo"),
   hubspot: () => import("./hubspot"),
-  "hubspot-revert": () => import("./hubspot-revert"),
   huddle01video: () => import("./huddle01video"),
   "ics-feedcalendar": () => import("./ics-feedcalendar"),
   jellyconferencing: () => import("./jelly"),
@@ -21,9 +20,7 @@ const appStore = {
   paypal: () => import("./paypal"),
   "pipedrive-crm": () => import("./pipedrive-crm"),
   salesforce: () => import("./salesforce"),
-  "salesforce-revert": () => import("./salesforce-revert"),
   zohocrm: () => import("./zohocrm"),
-  "zohocrm-revert": () => import("./zohocrm-revert"),
   sendgrid: () => import("./sendgrid"),
   stripepayment: () => import("./stripepayment"),
   tandemvideo: () => import("./tandemvideo"),
