+++ packages/app-store/pipedrive-crm/config.json
@@ -8,7 +8,7 @@
   "variant": "crm",
   "categories": ["crm"],
   "publisher": "Revert.dev ",
-  "email": "jatin@revert.dev",
+  "email": "support@revert.dev",
   "description": "Founded in 2010, Pipedrive is an easy and effective sales CRM that drives small business growth.\r\rToday, Pipedrive is used by revenue teams at more than 100,000 companies worldwide. Pipedrive is headquartered in New York and has offices across Europe and the US.\r\rThe company is backed by majority holder Vista Equity Partners, Bessemer Venture Partners, Insight Partners, Atomico, and DTCP.\r\rLearn more at www.pipedrive.com.",
   "isTemplate": false,
   "__createdUsingCli": true,
