--- packages/app-store/salesforce-revert/DESCRIPTION.md
@@ -1,13 +1,0 @@
----
-description: Salesforce (Sales Cloud) is a cloud-based application designed to help your salespeople sell smarter and faster by centralizing customer information, logging their interactions with your company, and automating many of the tasks salespeople do every day.
-items:
-  - 1.png
----
-
-Salesforce (Sales Cloud) is a cloud-based application designed to help your salespeople sell smarter and faster by centralizing customer information, logging their interactions with your company, and automating many of the tasks salespeople do every day.
-
-<br />
-<p class="mb-0">Features:</p>
-<ul class="mt-0">
-  <li>Creates event attendees as contacts in Salesforce (Sales Cloud)</li>
-</ul>
