--- packages/app-store/salesforce-revert/config.json
@@ -1,15 +1,0 @@
-{
-  "/*": "Don't modify slug - If required, do it using cli edit command",
-  "name": "Salesforce",
-  "slug": "salesforce-revert",
-  "type": "salesforce-revert_other_calendar",
-  "logo": "icon.png",
-  "url": "https://revert.dev/",
-  "variant": "other_calendar",
-  "categories": ["crm"],
-  "publisher": "Revert.dev",
-  "email": "jatin@revert.dev",
-  "description": "Salesforce (Sales Cloud) is a cloud-based application designed to help your salespeople sell smarter and faster by centralizing customer information, logging their interactions with your company, and automating many of the tasks salespeople do every day.",
-  "__createdUsingCli": true,
-  "dirName": "salesforce-revert"
-}
