--- packages/app-store/salesforce-revert/index.ts
@@ -1,2 +1,0 @@
-export * as api from "./api";
-export * as lib from "./lib";
