+++ packages/app-store/salesforce/config.json
@@ -4,11 +4,12 @@
   "slug": "salesforce",
   "type": "salesforce_crm",
   "logo": "icon.png",
-  "url": "https://cal.com/",
-  "variant": "crm",
+  "url": "https://revert.dev/",
+  "variant": "other_calendar",
   "categories": ["crm"],
-  "publisher": "Cal.com, Inc.",
-  "email": "help@cal.com",
+  "publisher": "Revert.dev",
+  "email": "support@revert.dev",
   "description": "Salesforce (Sales Cloud) is a cloud-based application designed to help your salespeople sell smarter and faster by centralizing customer information, logging their interactions with your company, and automating many of the tasks salespeople do every day.",
-  "__createdUsingCli": true
+  "__createdUsingCli": true,
+  "dirName": "salesforce"
 }
