--- packages/app-store/zohocrm-revert/DESCRIPTION.md
@@ -1,6 +1,0 @@
----
-items:
-  - 1.png
----
-
-{DESCRIPTION}
