--- packages/app-store/zohocrm-revert/config.json
@@ -1,17 +1,0 @@
-{
-  "/*": "Don't modify slug - If required, do it using cli edit command",
-  "name": "ZohoCRM",
-  "slug": "zohocrm-revert",
-  "type": "zohocrm-revert_other_calendar",
-  "logo": "icon.svg",
-  "url": "https://revert.dev",
-  "variant": "other",
-  "categories": ["crm"],
-  "publisher": "Revert.dev",
-  "email": "jatin@revert.dev",
-  "description": "Zoho CRM is a cloud-based application designed to help your salespeople sell smarter and faster by centralizing customer information, logging their interactions with your company, and automating many of the tasks salespeople do every day",
-  "isTemplate": false,
-  "__createdUsingCli": true,
-  "__template": "basic",
-  "dirName": "zohocrm-revert"
-}
