+++ packages/app-store/zohocrm/config.json
@@ -4,13 +4,14 @@
   "slug": "zohocrm",
   "type": "zohocrm_crm",
   "logo": "icon.svg",
-  "url": "https://github.com/jatinsandilya",
-  "variant": "crm",
+  "url": "https://revert.dev",
+  "variant": "other",
   "categories": ["crm"],
-  "publisher": "Jatin Sandilya",
-  "email": "help@cal.com",
+  "publisher": "Revert.dev",
+  "email": "support@revert.dev",
   "description": "Zoho CRM is a cloud-based application designed to help your salespeople sell smarter and faster by centralizing customer information, logging their interactions with your company, and automating many of the tasks salespeople do every day",
   "isTemplate": false,
   "__createdUsingCli": true,
-  "__template": "basic"
+  "__template": "basic",
+  "dirName": "zohocrm"
 }
